client/requirements.txt
PySide6-Essentials>=6.8,<7.0
cryptography>=43,<46
argon2-cffi>=25,<26
httpx>=0.27,<1.0
